+++ devkit/samples/pending-exception/pending-exception.ts
@@ -1,0 +1,5 @@
+import { Given, PendingException } from '@cucumber/fake-cucumber'
+
+Given('an unimplemented pending step', function () {
+  throw new PendingException('TODO')
+})