+++ devkit/samples/skipped-exception/skipped-exception.ts
@@ -1,0 +1,5 @@
+import { Given, SkippedException } from '@cucumber/fake-cucumber'
+
+Given('I skip a step', function () {
+  throw new SkippedException('skipping')
+})