+++ devkit/src/PendingException.ts
@@ -1,0 +1,1 @@
+export class PendingException extends Error {}