+++ devkit/src/SkippedException.ts
@@ -1,0 +1,1 @@
+export class SkippedException extends Error {}